Add a new request decider event loop
