+++ signer/src/lib.rs
@@ -21,6 +21,7 @@ pub mod logging;
 pub mod message;
 pub mod network;
 pub mod proto;
+pub mod request_decider;
 pub mod signature;
 pub mod stacks;
 pub mod storage;
