+++ signer/src/testing/mod.rs
@@ -9,6 +9,7 @@ pub mod context;
 pub mod dummy;
 pub mod message;
 pub mod network;
+pub mod request_decider;
 pub mod stacks;
 pub mod storage;
 pub mod transaction_coordinator;
