+++ signer/tests/integration/request_decider.rs
@@ -1,0 +1,284 @@
+use std::sync::atomic::Ordering;
+
+use rand::SeedableRng as _;
+
+use signer::bitcoin::MockBitcoinInteract;
+use signer::context::Context;
+use signer::emily_client::MockEmilyInteract;
+use signer::keys::PrivateKey;
+use signer::network::InMemoryNetwork;
+use signer::request_decider::RequestDeciderEventLoop;
+use signer::stacks::api::MockStacksInteract;
+use signer::storage::model::BitcoinBlockHash;
+use signer::storage::postgres::PgStore;
+use signer::storage::DbRead as _;
+use signer::testing;
+use signer::testing::context::*;
+use signer::testing::request_decider::TestEnvironment;
+
+use crate::setup::backfill_bitcoin_blocks;
+use crate::setup::TestSweepSetup;
+use crate::DATABASE_NUM;
+
+fn test_environment(
+    db: PgStore,
+    signing_threshold: u32,
+    num_signers: usize,
+) -> TestEnvironment<
+    TestContext<
+        PgStore,
+        WrappedMock<MockBitcoinInteract>,
+        WrappedMock<MockStacksInteract>,
+        WrappedMock<MockEmilyInteract>,
+    >,
+> {
+    let context_window = 6;
+
+    let test_model_parameters = testing::storage::model::Params {
+        num_bitcoin_blocks: 20,
+        num_stacks_blocks_per_bitcoin_block: 3,
+        num_deposit_requests_per_block: 5,
+        num_withdraw_requests_per_block: 5,
+        num_signers_per_request: 0,
+    };
+
+    let context = TestContext::builder()
+        .with_storage(db)
+        .with_mocked_clients()
+        .build();
+
+    TestEnvironment {
+        context,
+        num_signers,
+        context_window,
+        signing_threshold,
+        test_model_parameters,
+    }
+}
+
+async fn create_signer_database() -> PgStore {
+    let db_num = DATABASE_NUM.fetch_add(1, Ordering::SeqCst);
+    signer::testing::storage::new_test_database(db_num, true).await
+}
+
+#[cfg_attr(not(feature = "integration-tests"), ignore)]
+#[tokio::test]
+async fn should_store_decisions_for_pending_deposit_requests() {
+    let num_signers = 3;
+    let signing_threshold = 2;
+
+    let db = create_signer_database().await;
+    // We need to clone the connection so that we can drop the associated
+    // databases later.
+    test_environment(db.clone(), signing_threshold, num_signers)
+        .assert_should_store_decisions_for_pending_deposit_requests()
+        .await;
+
+    // Now drop the database that we just created.
+    signer::testing::storage::drop_db(db).await;
+}
+
+#[cfg_attr(not(feature = "integration-tests"), ignore)]
+#[tokio::test]
+async fn should_store_decisions_for_pending_withdraw_requests() {
+    let num_signers = 3;
+    let signing_threshold = 2;
+
+    let db = create_signer_database().await;
+    // We need to clone the connection so that we can drop the associated
+    // databases later.
+    test_environment(db.clone(), signing_threshold, num_signers)
+        .assert_should_store_decisions_for_pending_withdrawal_requests()
+        .await;
+
+    // Now drop the database that we just created.
+    signer::testing::storage::drop_db(db).await;
+}
+
+#[cfg_attr(not(feature = "integration-tests"), ignore)]
+#[tokio::test]
+async fn should_store_decisions_received_from_other_signers() {
+    let num_signers = 3;
+    let signing_threshold = 2;
+
+    let db = create_signer_database().await;
+    // We need to clone the connection so that we can drop the associated
+    // databases later.
+    test_environment(db.clone(), signing_threshold, num_signers)
+        .assert_should_store_decisions_received_from_other_signers()
+        .await;
+
+    // Now drop the database that we just created.
+    signer::testing::storage::drop_db(db).await;
+}
+
+/// Test that [`TxSignerEventLoop::handle_pending_deposit_request`] does
+/// not error when attempting to check the scriptPubKeys of the
+/// inputs of a deposit.
+#[cfg_attr(not(feature = "integration-tests"), ignore)]
+#[tokio::test]
+async fn handle_pending_deposit_request_address_script_pub_key() {
+    let db_num = DATABASE_NUM.fetch_add(1, Ordering::SeqCst);
+    let db = testing::storage::new_test_database(db_num, true).await;
+
+    let mut rng = rand::rngs::StdRng::seed_from_u64(51);
+
+    let ctx = TestContext::builder()
+        .with_storage(db.clone())
+        .with_mocked_clients()
+        .build();
+
+    let (rpc, faucet) = sbtc::testing::regtest::initialize_blockchain();
+
+    // This confirms a deposit transaction, and has a nice helper function
+    // for storing a real deposit.
+    let setup = TestSweepSetup::new_setup(rpc, faucet, 10000, &mut rng);
+
+    // Let's get the blockchain data into the database.
+    let chain_tip: BitcoinBlockHash = setup.sweep_block_hash.into();
+    backfill_bitcoin_blocks(&db, rpc, &chain_tip).await;
+
+    // We need to store the deposit request because of the foreign key
+    // constraint on the deposit_signers table.
+    setup.store_deposit_request(&db).await;
+
+    // In order to fetch the deposit request that we just store, we need to
+    // store the deposit transaction.
+    setup.store_deposit_tx(&db).await;
+
+    // When we run TxSignerEventLoop::handle_pending_deposit_request, we
+    // check if the current signer is in the signing set. For this check we
+    // need a row in the dkg_shares table.
+    setup.store_dkg_shares(&db).await;
+
+    let mut requests = db
+        .get_pending_deposit_requests(&chain_tip, 100)
+        .await
+        .unwrap();
+    // There should only be the one deposit request that we just fetched.
+    assert_eq!(requests.len(), 1);
+    let request = requests.pop().unwrap();
+
+    let network = InMemoryNetwork::new();
+    let mut tx_signer = RequestDeciderEventLoop {
+        network: network.connect(),
+        context: ctx.clone(),
+        context_window: 10000,
+        blocklist_checker: Some(()),
+        signer_private_key: setup.aggregated_signer.keypair.secret_key().into(),
+    };
+
+    // We need this so that there is a live "network". Otherwise,
+    // TxSignerEventLoop::handle_pending_deposit_request will error when
+    // trying to send a message at the end.
+    let _rec = ctx.get_signal_receiver();
+
+    // We don't want this to error. There was a bug before, see
+    // https://github.com/stacks-network/sbtc/issues/674.
+    tx_signer
+        .handle_pending_deposit_request(request, &chain_tip)
+        .await
+        .unwrap();
+
+    // A decision should get stored and there should only be one
+    let outpoint = setup.deposit_request.outpoint;
+    let mut votes = db
+        .get_deposit_signers(&outpoint.txid.into(), outpoint.vout)
+        .await
+        .unwrap();
+    assert_eq!(votes.len(), 1);
+
+    // The blocklist checker that we have configured accepts all deposits.
+    // Also we are in the signing set so we can sign for the deposit.
+    let vote = votes.pop().unwrap();
+    assert!(vote.can_sign);
+    assert!(vote.can_accept);
+
+    testing::storage::drop_db(db).await;
+}
+
+/// Test that [`TxSignerEventLoop::handle_pending_deposit_request`] will
+/// write the can_sign field to be false if the current signer is not part
+/// of the signing set locking the deposit transaction.
+#[cfg_attr(not(feature = "integration-tests"), ignore)]
+#[tokio::test]
+async fn handle_pending_deposit_request_not_in_signing_set() {
+    let db_num = DATABASE_NUM.fetch_add(1, Ordering::SeqCst);
+    let db = testing::storage::new_test_database(db_num, true).await;
+
+    let mut rng = rand::rngs::StdRng::seed_from_u64(51);
+
+    let ctx = TestContext::builder()
+        .with_storage(db.clone())
+        .with_mocked_clients()
+        .build();
+
+    let (rpc, faucet) = sbtc::testing::regtest::initialize_blockchain();
+
+    // This confirms a deposit transaction, and has a nice helper function
+    // for storing a real deposit.
+    let setup = TestSweepSetup::new_setup(rpc, faucet, 10000, &mut rng);
+
+    // Let's get the blockchain data into the database.
+    let chain_tip: BitcoinBlockHash = setup.sweep_block_hash.into();
+    backfill_bitcoin_blocks(&db, rpc, &chain_tip).await;
+
+    // We need to store the deposit request because of the foreign key
+    // constraint on the deposit_signers table.
+    setup.store_deposit_request(&db).await;
+
+    // In order to fetch the deposit request that we just store, we need to
+    // store the deposit transaction.
+    setup.store_deposit_tx(&db).await;
+
+    // When we run TxSignerEventLoop::handle_pending_deposit_request, we
+    // check if the current signer is in the signing set and this adds a
+    // signing set.
+    setup.store_dkg_shares(&db).await;
+
+    let mut requests = db
+        .get_pending_deposit_requests(&chain_tip, 100)
+        .await
+        .unwrap();
+    // There should only be the one deposit request that we just fetched.
+    assert_eq!(requests.len(), 1);
+    let request = requests.pop().unwrap();
+
+    let network = InMemoryNetwork::new();
+    let mut tx_signer = RequestDeciderEventLoop {
+        network: network.connect(),
+        context: ctx.clone(),
+        context_window: 10000,
+        blocklist_checker: Some(()),
+        // We generate a new private key here so that we know (with very
+        // high probability) that this signer is not in the signer set.
+        signer_private_key: PrivateKey::new(&mut rng),
+    };
+
+    // We need this so that there is a live "network". Otherwise,
+    // TxSignerEventLoop::handle_pending_deposit_request will error when
+    // trying to send a message at the end.
+    let _rec = ctx.get_signal_receiver();
+
+    tx_signer
+        .handle_pending_deposit_request(request, &chain_tip)
+        .await
+        .unwrap();
+
+    // A decision should get stored and there should only be one
+    let outpoint = setup.deposit_request.outpoint;
+    let mut votes = db
+        .get_deposit_signers(&outpoint.txid.into(), outpoint.vout)
+        .await
+        .unwrap();
+    assert_eq!(votes.len(), 1);
+
+    // can_sign should be false since the public key associated with our
+    // random private key is not in the signing set. And can_accept is
+    // always true with the given blocklist client.
+    let vote = votes.pop().unwrap();
+    assert!(!vote.can_sign);
+    assert!(vote.can_accept);
+
+    testing::storage::drop_db(db).await;
+}
